Close the webdriver connection after the last test
Previously, we waited for the object to be destructed before closing
the webdriver connection. As the object only gets destructed at the
end of the full test run, this meant that each scene caused one
connection to remain open, until we hit the limit of the hub, at
which point the test failed.